requirements.txt
pillow
platformdirs
html2image